+++ extension-library/src/main/java/app/morphe/extension/shared/Logger.java
@@ -11,18 +11,78 @@ import androidx.annotation.Nullable;
 
 import java.io.PrintWriter;
 import java.io.StringWriter;
+import java.util.Collections;
+import java.util.Deque;
+import java.util.Objects;
+import java.util.SequencedCollection;
+import java.util.concurrent.ConcurrentLinkedDeque;
+import java.util.concurrent.atomic.AtomicInteger;
 
 import app.morphe.extension.shared.settings.BaseSettings;
-import app.morphe.extension.shared.settings.preference.LogBufferManager;
 
 /**
  * Morphe specific logger. Logging is done to standard device log (accessible through ADB),
- * and additionally accessible through {@link LogBufferManager}.
+ * and additionally accessible through {@link #getLogBuffer()}.
  *
  * All methods are thread safe, and are safe to call even
  * if {@link Utils#getContext()} is not available.
  */
 public class Logger {
+
+    /** Maximum byte size of all buffer entries. Must be less than Android's 1 MB Binder transaction limit. */
+    private static final int BUFFER_MAX_BYTES = 900_000;
+    /** Limit number of log lines. */
+    private static final int BUFFER_MAX_SIZE = 10_000;
+
+    private static final Deque<String> logBuffer = new ConcurrentLinkedDeque<>();
+    private static final AtomicInteger logBufferByteSize = new AtomicInteger();
+
+    public static String getFilteredLogs() {
+        StringBuilder filteredOutput = new StringBuilder(logBufferByteSize.get());
+
+        for (String log : logBuffer) {
+            filteredOutput.append(log).append('\n');
+        }
+
+        return filteredOutput.toString().trim();
+    }
+
+    private static void appendToLogBuffer(String message) {
+        Objects.requireNonNull(message);
+
+        // It's very important that no Settings are used in this method,
+        // as this code is used when a context is not set and thus referencing
+        // a setting will crash the app.
+        logBuffer.addLast(message);
+        int newSize = logBufferByteSize.addAndGet(message.length());
+
+        // Remove the oldest entries if over the log size limits.
+        while (newSize > BUFFER_MAX_BYTES || logBuffer.size() > BUFFER_MAX_SIZE) {
+            String removed = logBuffer.pollFirst();
+            if (removed == null) {
+                // Thread race of two different calls to this method, and the other thread won.
+                return;
+            }
+
+            newSize = logBufferByteSize.addAndGet(-removed.length());
+        }
+    }
+
+    public static Deque<String> getLogBuffer() {
+        return logBuffer;
+    }
+
+    public static void clearLogBufferData() {
+        // Cannot simply clear the log buffer because there is no
+        // write lock for both the deque and the atomic int.
+        // Instead, pop off log entries and decrement the size one by one.
+        while (!logBuffer.isEmpty()) {
+            String removed = logBuffer.pollFirst();
+            if (removed != null) {
+                logBufferByteSize.addAndGet(-removed.length());
+            }
+        }
+    }
 
     /**
      * Log messages using lambdas.
@@ -76,9 +136,9 @@ public class Logger {
     }
 
     /**
-     * Internal method to handle logging to Android Log and {@link LogBufferManager}.
-     * Appends the log message, stack trace (if enabled), and exception (if present) to logBuffer
-     * with class name but without 'morphe:' prefix.
+     * Internal method to handle logging to Android Log and internally to appends the log message,
+     * stack trace (if enabled), and exception (if present) to logBuffer with class name but
+     * without 'morphe:' prefix.
      *
      * @param logLevel          The log level.
      * @param message           Log message object.
@@ -116,7 +176,7 @@ public class Logger {
 
         // Do not include "morphe:" prefix in clipboard logs.
         String managerToastString = className + ": " + logText;
-        LogBufferManager.appendToLogBuffer(managerToastString);
+        appendToLogBuffer(managerToastString);
 
         String logTag = MORPHE_LOG_TAG_PREFIX + className;
         switch (logLevel) {